refactor(extensions): change module name and artifactId from blob-core to azure-blob-core
