+++ extensions/common/azure/azure-blob-core/src/main/java/org/eclipse/dataspaceconnector/azure/blob/core/adapter/BlobAdapter.java
