+++ extensions/common/azure/azure-blob-core/src/main/resources/META-INF/services/org.eclipse.dataspaceconnector.spi.system.ServiceExtension
