+++ extensions/common/azure/azure-blob-core/src/testFixtures/java/org/eclipse/dataspaceconnector/azure/blob/core/AzureStorageTestFixtures.java
