+++ extensions/common/azure/azure-resource-manager/build.gradle.kts
@@ -20,7 +20,7 @@ plugins {
 }
 
 dependencies {
-    implementation(project(":extensions:common:azure:blob-core"))
+    implementation(project(":extensions:common:azure:azure-blob-core"))
     implementation(project(":core:common:util"))
     implementation("com.azure:azure-identity:${azureIdentityVersion}")
     implementation("com.azure.resourcemanager:azure-resourcemanager:${azureResourceManagerVersion}")
