+++ settings.gradle.kts
@@ -69,7 +69,7 @@ include(":extensions:common:aws:aws-s3-core")
 include(":extensions:common:azure:azure-eventgrid")
 include(":extensions:common:azure:azure-resource-manager")
 include(":extensions:common:azure:azure-test")
-include(":extensions:common:azure:blob-core")
+include(":extensions:common:azure:azure-blob-core")
 include(":extensions:common:azure:cosmos-common")
 include(":extensions:common:configuration:filesystem-configuration")
 include(":extensions:common:events:cloudevents-http")
