+++ system-tests/azure-data-factory-tests/build.gradle.kts
@@ -36,7 +36,7 @@ dependencies {
         exclude(group = "io.gatling", module = "gatling-graphite")
     }
 
-    testImplementation(project(":extensions:common:azure:blob-core"))
+    testImplementation(project(":extensions:common:azure:azure-blob-core"))
     testImplementation(project(":extensions:common:vault:azure-vault"))
     testImplementation(project(":extensions:common:junit"))
     testImplementation(testFixtures(project(":system-tests:tests")))
@@ -50,4 +50,3 @@ dependencies {
     testRuntimeOnly(project(":system-tests:runtimes:azure-data-factory-transfer-provider"))
     testRuntimeOnly(project(":system-tests:runtimes:azure-data-factory-transfer-consumer"))
 }
-
